Revert "[Internal Manager] Implement internal balance withdraw flow (#74)"
This reverts commit 5527d46965406a5c84c0aae338bd7809b8068f60.
