+++ packages/gql/src/balancer-internal-manager/internalBalances.ts
@@ -4,26 +4,8 @@ export const internalBalance = gql`
   query InternalBalance($userAddress: ID!) {
     user(id: $userAddress) {
       id
-      userInternalBalances(where: { balance_gt: 0 }) {
+      userInternalBalances {
         tokenInfo {
-          name
-          symbol
-          address
-          decimals
-        }
-        balance
-      }
-    }
-  }
-`;
-
-export const singleInternalBalance = gql`
-  query SingleInternalBalance($userAddress: ID!, $tokenAddress: Bytes!) {
-    user(id: $userAddress) {
-      id
-      userInternalBalances(where: { token: $tokenAddress }) {
-        tokenInfo {
-          name
           symbol
           address
           decimals
