--- packages/next-app/src/app/internalmanager/[network]/[operationKind]/layout.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-export default function Layout({ children }: React.PropsWithChildren) {
-  return children;
-}
