+++ packages/next-app/src/app/internalmanager/layout.tsx
@@ -3,8 +3,8 @@ import * as React from "react";
 
 import balancerSymbol from "#/assets/balancer-symbol.svg";
 import { Header } from "#/components/Header";
-import { InternalManagerProvider } from "#/contexts/InternalManagerContext";
-import { NetworksContextProvider } from "#/contexts/networks";
+
+import { InternalManagerProvider } from "./(components)/InternalManagerProvider";
 
 export const metadata: Metadata = {
   title: "Balancer Internal Manager",
@@ -13,15 +13,13 @@ export const metadata: Metadata = {
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
-    <NetworksContextProvider>
-      <div className="flex h-screen flex-col">
-        <Header
-          linkUrl={"/internalmanager"}
-          title={"Internal Manager"}
-          imageSrc={balancerSymbol}
-        />
-        <InternalManagerProvider>{children}</InternalManagerProvider>
-      </div>
-    </NetworksContextProvider>
+    <div className="flex h-screen flex-col">
+      <Header
+        linkUrl={"/internalmanager"}
+        title={"Internal Manager"}
+        imageSrc={balancerSymbol}
+      />
+      <InternalManagerProvider>{children}</InternalManagerProvider>
+    </div>
   );
 }
