+++ packages/next-app/src/hooks/useTransaction.ts
@@ -1,11 +1,8 @@
 import { InternalBalanceQuery } from "@balancer-pool-metadata/gql/src/balancer-internal-manager/__generated__/Mainnet";
 import { getNetworkUrl } from "@balancer-pool-metadata/shared";
 import { parseFixed } from "@ethersproject/bignumber";
-import { useRouter } from "next/navigation";
 import { Dispatch, useEffect, useState } from "react";
-import { FieldValues } from "react-hook-form";
 
-import { useInternalBalance } from "#/contexts/InternalManagerContext";
 import { PoolMetadataAttribute } from "#/contexts/PoolMetadataContext";
 import { UserBalanceOpKind } from "#/lib/internal-balance-helper";
 import { pinJSON } from "#/lib/ipfs";
@@ -36,16 +33,10 @@ enum NotificationVariant {
   SUCCESS = "success",
 }
 
-export type Notification = {
+type Notification = {
   title: string;
   description: string;
   variant: NotificationVariant;
-};
-
-type SubmitData = {
-  receiverAddress: `0x${string}`;
-  tokenAddress: `0x${string}`;
-  tokenAmount: string;
 };
 
 type TransactionHookResult = {
@@ -226,34 +217,23 @@ export function useMetadataTransaction({
 export function useInternalBalancesTransaction({
   userAddress,
   token,
-  operationKind,
 }: {
   userAddress: `0x${string}`;
   token: ArrElement<GetDeepProp<InternalBalanceQuery, "userInternalBalances">>;
-  operationKind: UserBalanceOpKind | null;
 }) {
-  const {
-    setNotification,
-    setTransactionUrl,
-    isNotifierOpen,
-    setIsNotifierOpen,
-    notification,
-  } = useInternalBalance();
-  const { push } = useRouter();
   const { chain } = useNetwork();
-  const [submitData, setSubmitData] = useState<SubmitData | null>(null);
+  const [isNotifierOpen, setIsNotifierOpen] = useState(false);
+  const [notification, setNotification] = useState<Notification | null>(null);
+  const [transactionUrl, setTransactionUrl] = useState<string>();
+  const [operationKind, setOperationKind] = useState<UserBalanceOpKind>();
 
   //Prepare data for transaction
   const userBalanceOp = {
     kind: operationKind as number,
-    asset: submitData?.tokenAddress as `0x${string}`,
-    //TODO get this if tokenAmount is not defined a better solution than 0 to initialize the value
-    amount: parseFixed(
-      submitData?.tokenAmount ? submitData.tokenAmount : "0",
-      token.tokenInfo?.decimals
-    ),
+    asset: token.tokenInfo.address as `0x${string}`,
+    amount: parseFixed(token.balance, token.tokenInfo.decimals),
     sender: userAddress as `0x${string}`,
-    recipient: submitData?.receiverAddress as `0x${string}`,
+    recipient: userAddress as `0x${string}`,
   };
 
   const { config } = usePrepareVaultManageUserBalance({
@@ -263,22 +243,19 @@ export function useInternalBalancesTransaction({
   const { data, write } = useVaultManageUserBalance(config);
 
   //trigger transaction
-  function handleWithdraw(data: FieldValues) {
-    setSubmitData({
-      tokenAddress: data.tokenAddress,
-      tokenAmount: data.tokenAmount,
-      receiverAddress: data.receiverAddress,
-    });
-  }
-
-  // //trigger the actual transaction
-  useEffect(() => {
-    if (!submitData) return;
+  function handleWithdraw() {
+    setOperationKind(UserBalanceOpKind.WITHDRAW_INTERNAL);
     setNotification(
       NOTIFICATION_MAP_INTERNAL_BALANCES[TransactionStatus.WAITING_APPROVAL]
     );
+  }
+
+  //trigger the actual transaction
+  useEffect(() => {
+    if (!operationKind) return;
     write?.();
-  }, [submitData]);
+    setOperationKind(undefined);
+  }, [operationKind]);
 
   //handle transaction status
   useEffect(() => {
@@ -299,13 +276,11 @@ export function useInternalBalancesTransaction({
   useWaitForTransaction({
     hash: data?.hash,
     onSuccess() {
-      push(`/internalmanager`);
       setNotification(
         NOTIFICATION_MAP_INTERNAL_BALANCES[TransactionStatus.CONFIRMED]
       );
     },
     onError() {
-      push(`/internalmanager`);
       setNotification(
         NOTIFICATION_MAP_INTERNAL_BALANCES[TransactionStatus.WRITE_ERROR]
       );
@@ -329,6 +304,11 @@ export function useInternalBalancesTransaction({
   }, [notification]);
 
   return {
+    isNotifierOpen,
+    setIsNotifierOpen,
+    operationKind,
     handleWithdraw,
+    notification,
+    transactionUrl,
   };
 }
